chore(anyharness): split oversized portability/mobility/route-auth test modules
check_max_lines flagged three files over the 600-line cap after the mobility
migration work:

- portability/mod.rs (940 -> 491): move the #[cfg(test)] mod tests body to a
  sibling portability/tests.rs (declared 'mod tests;'); 'use super::*' still
  resolves to the portability module.
- route_auth/render_tests.rs (607 -> 577): move the codex api_key render test
  this PR added into route_auth/render_codex_tests.rs (declared alongside
  'mod render_tests;'), self-contained with its own minimal resolver + state
  builders.
- mobility/service.rs (2180 -> 1494): extract the #[cfg(test)] mod tests block
  into service_tests.rs (validation/preflight) and service_install_tests.rs
  (install/round-trip), both compiled as children of 'service' via #[path] so
  'use super::*' keeps seeing service's private items. Allowlist lowered
  1520 -> 1494 (shrink, not a bump).

Every resulting file is under 600 (service.rs under its 1494 allowlist).
